Move retrieval machinery to network/retrieval and revocation to policy/revocation
